+++ server/src/test/java/com/zzuiksa/server/domain/schedule/entity/ScheduleTests.java
@@ -285,8 +285,8 @@ public class ScheduleTests {
     public void setToPlace_nameAndLatLng_success() {
         // given
         String toPlaceName = "Seoul";
-        float lat = 123.4567f;
-        float lng = 35.79f;
+        double lat = 123.4567;
+        double lng = 35.79;
         Schedule schedule = new Schedule();
 
         // when
@@ -356,8 +356,8 @@ public class ScheduleTests {
     public void setFromPlace_nameAndLatLng_success() {
         // given
         String fromPlaceName = "Seoul";
-        float lat = 123.4567f;
-        float lng = 35.79f;
+        double lat = 123.4567;
+        double lng = 35.79;
         Schedule schedule = new Schedule();
 
         // when
